LPD-101794 Move the Custom Element tests to their own project
Every spec in client-extension-web.main pays the deploy cost of all the sample client extensions the project lists, whatever it actually asserts on. The Custom Element specs need liferay-sample-custom-element-1, 3, 4, and 5, plus liferay-sample-etc-frontend for its custom element.

They now run under client-extension-web.custom-element, whose env folder lists only those. The parent env folders still apply, since common.sh walks up from the project directory to the Playwright base directory and applies every ancestor, so the project needs no portal-ext.properties of its own. The only property main adds on top of the base one is feature.flag.LPS-164563, which gates the FDS Cell Renderer and FDS Filter client extension types and is therefore irrelevant here.

The editCustomElementPageTest fixture and the EditCustomElementPage class move with the specs. Both are also used from client-extension-web.main and users-admin-web.main, whose imports are repointed at the new location.

liferay-sample-custom-element-2 is dropped rather than moved. No spec references it, so main was deploying it for nothing, and carrying it into the new project would defeat the purpose of the split.

The project keeps reporting to the Frontend Client Extension component and joins the same batches as main, since the Custom Element tests stay with the Frontend Infrastructure team.
