+++ modules/apps/client-extension/client-extension-web/test.properties
@@ -3,6 +3,7 @@ modified.files.includes[relevant][client-extension-web-playwright-rule]=**
 playwright.projects.includes[playwright-js-tomcat101-postgresql163][relevant][client-extension-web-playwright-rule]=\
     client-extension-web.cluster,\
     client-extension-web.css,\
+    client-extension-web.custom-element,\
     client-extension-web.js,\
     client-extension-web.main,\
     client-extension-web.theme-css,\
