+++ modules/test/playwright/tests/client-extension-web/custom-element/config.ts
@@ -1,0 +1,12 @@
+/**
+ * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
+ * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
+ */
+
+export const config = {
+	name: 'client-extension-web.custom-element',
+	testDir: 'tests/client-extension-web/custom-element',
+	use: {
+		testIdAttribute: 'data-qa-id',
+	},
+};
