+++ modules/test/playwright/tests/client-extension-web/custom-element/customElement.spec.ts
