+++ modules/test/playwright/tests/client-extension-web/custom-element/test.properties
@@ -1,0 +1,1 @@
+testray.main.component.name=Frontend Client Extension
