+++ modules/test/playwright/tests/client-extension-web/main/clientExtensionImportExport.spec.ts
@@ -8,9 +8,9 @@ import path from 'path';
 
 import {loginTest} from '../../../fixtures/loginTest';
 import getRandomString from '../../../utils/getRandomString';
+import {editCustomElementPageTest} from '../custom-element/fixtures/editCustomElementPageTest';
 import {clientExtensionsPageTest} from '../fixtures/clientExtensionsPageTest';
 import {WaitAction} from '../pages/EditClientExtensionsPage';
-import {editCustomElementPageTest} from './fixtures/editCustomElementPageTest';
 
 const test = mergeTests(
 	clientExtensionsPageTest,
