+++ test.properties
@@ -3951,6 +3951,7 @@
 
     playwright.projects.includes[playwright-js-tomcat101-*][frontend]=\
         client-extension-web.cluster,\
+        client-extension-web.custom-element,\
         client-extension-web.main,\
         cookies-banner-web.main,\
         frontend-data-set-admin-web.main,\
@@ -4006,6 +4007,7 @@
 
     playwright.projects.includes[playwright-js-tomcat101-*][frontend-client-extensions]=\
         client-extension-web.cluster,\
+        client-extension-web.custom-element,\
         client-extension-web.main
 
     test.batch.class.names.includes[frontend-client-extensions]=\
